Switch to bin/execute.rs
